Change the default value of `org-mime-instructions-hint`
Currently, the default value of `org-mime-instructions-hint` begins
with two pound signs ("##").  However, two pound signs are not
recognized by org-mode as comments.  Therefore, when editing email
contents in the `org-mime-src-mode` org buffer, the hint is not
fontified as a comment. 

Additionally, if the user mistakenly inserts text prior to the hint,
the hint text will not be removed by a later call to
`org-mime-edit-src-exit` or `org-mime-edit-src-save`, resulting in the
text being present in the mail composition buffer.

Therefore, turn the hint into an org comment.